fixed dmg and ability. sort of

+++ Assets/Scripts/Test_Scripts/Hugo_TestScripts/Test_Enemy.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int health = 50;
+    public float health = 50;
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         health -= damage;
         Debug.Log("Enemy took " + damage + " damage!");
